Move `ClassDefinition` content to `ClassDefinitionTerm`
